Fix benchmark tests (#283)
* Minimum requirement is now 14.17.6

* Check if repo exists from previous run

* Add comment

* Reduce to 1GB because GitHub runner are running out of mem

+++ benchmark/snowfs-vs-git.ts
@@ -9,7 +9,7 @@ import { Repository, RESET } from '../src/repository';
 // eslint-disable-next-line import/no-extraneous-dependencies
 const chalk = require('chalk');
 
-const BENCHMARK_FILE_SIZE = 4000000000;
+const BENCHMARK_FILE_SIZE = process.platform === 'win32' ? 1000000000 : 4000000000;
 
 async function input(question: string): Promise<string> {
   const rl = readline.createInterface({
@@ -100,7 +100,9 @@ async function createFile(dst: string, size: number, t = console) {
 }
 
 async function gitAddTexture(repoPath: string, textureFilesize: number = BENCHMARK_FILE_SIZE, t = console): Promise<number> {
-  fse.rmdirSync(repoPath, { recursive: true });
+  if (fse.pathExistsSync(repoPath)) {
+    fse.rmdirSync(repoPath, { recursive: true });
+  }
 
   t.log(`Create Git(+LFS) Repository at: ${repoPath}`);
   await exec('git', ['init', basename(repoPath)], t, { cwd: dirname(repoPath) });
@@ -200,6 +202,7 @@ export async function startBenchmark(textureFilesize: number = BENCHMARK_FILE_SI
       playground = os.tmpdir();
     }
 
+    // Test if we can create directories in the playground
     const tmp: string = join(playground, 'benchmark-xyz-test');
     try {
       fse.mkdirpSync(tmp);
